Make Http1xPool and ClientHandler not anymore inner classes of ConnectionManager like Http2Pool
